\name{astrolibR-package}
\alias{astrolibR-package}
\alias{astrolibR}
\docType{package}
\title{astrolibR: Astronomy Users Library}
\description{Several dozen low-level utilities and codes from the Interactive Data Language (IDL) Astronomy Users Library (http://idlastro.gsfc.nasa.gov) are implemented in R.  They treat: time, coordinate and proper motion transformations; terrestrial precession and nutation, atmospheric refraction and aberration, barycentric corrections, and related effects; utilities for astrometry, photometry, and spectroscopy; and utilities for planetary, stellar, Galactic, and extragalactic science.}
\details{
\tabular{ll}{
Package: \tab astrolibR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-30\cr
License: \tab GPL\cr
}
}
