\name{intdiv}
\alias{intdiv}

\title{
Integer divide
}
\description{
Coerces dividend and divisor into integer prior to divide
}
\usage{
intdiv(dividend, divisor)
}
\arguments{
  \item{dividend}{scalar or vector}
  \item{divisor}{scalar or vector}
}
\author{
Arnab Chakraborty    2013}
\examples{
intdiv(5.6,-3.1)
intdiv(6.6,-3.1)
}
\keyword{ misc }
